Add `polonius_facts` crate containing `AllFacts`
This breaks the cyclical dependency between `polonius-engine` and
`polonius-souffle`.
